字体都改为彩色 按钮弄下 数据取两位小数 数据显示框弄紧凑点！

+++ files/ScaleOfCapitalFlows.sql
@@ -13,21 +13,21 @@ create table SOCF_CUDDINGTON (
     ID int primary key auto_increment,
     NIAN_JD varchar(20) unique,
     is_nian_jd boolean,
-    S1 double(30,4),
-    S2 double(30,4),
-    S3 double(30,4),
-    S4 double(30,4),
-    S5 double(30,4),
-    S6 double(30,4),
-    S7 double(30,4),
-    S8 double(30,4),
-    S9 double(30,4),
-    S10 double(30,4),
-    S11 double(30,4),
-    S12 double(30,4),
-    S13 double(30,4),
-    S14 double(30,4),
-    S15 double(30,4)
+    S1 double(30,2),
+    S2 double(30,2),
+    S3 double(30,2),
+    S4 double(30,2),
+    S5 double(30,2),
+    S6 double(30,2),
+    S7 double(30,2),
+    S8 double(30,2),
+    S9 double(30,2),
+    S10 double(30,2),
+    S11 double(30,2),
+    S12 double(30,2),
+    S13 double(30,2),
+    S14 double(30,2),
+    S15 double(30,2)
 );
 
 #间接法（Residual method）：
@@ -35,23 +35,23 @@ create table SOCF_CUDDINGTON (
 create table SOCF_RESIDUAL(
     ID int primary key auto_increment,
     NIAN_JD varchar(20) unique,
-    S double(30,4),
-    L1 double(30,4),
-    L2 double(30,4),
-    L3 double(30,4),
-    L4 double(30,4),
-    L5 double(30,4),
-    L6 double(30,4),
-    L7 double(30,4),
-    L8 double(30,4),
-    L9 double(30,4),
-    L10 double(30,4),
-    L11 double(30,4),
-    L12 double(30,4),
-    L13 double(30,4),
-    L14 double(30,4),
-    L15 double(30,4),
-    L16 double(30,4),
+    S double(30,2),
+    L1 double(30,2),
+    L2 double(30,2),
+    L3 double(30,2),
+    L4 double(30,2),
+    L5 double(30,2),
+    L6 double(30,2),
+    L7 double(30,2),
+    L8 double(30,2),
+    L9 double(30,2),
+    L10 double(30,2),
+    L11 double(30,2),
+    L12 double(30,2),
+    L13 double(30,2),
+    L14 double(30,2),
+    L15 double(30,2),
+    L16 double(30,2),
     is_nian_jd boolean
 );
 
@@ -77,9 +77,9 @@ create table t_user(
 create table SOCF_YD_RESIDUAL (
     id int primary key auto_increment,
     nian_yd varchar(20) not null unique,
-    s double(30,4),
-    fdi double(30,4),
-    fbt double(30,4)
+    s double(30,2),
+    fdi double(30,2),
+    fbt double(30,2)
 );
 
 #MONTHSCALE
